changed to doubles

+++ app/src/main/java/bfst21/osm/Node.java
@@ -13,9 +13,9 @@ import bfst21.Model;
 public class Node extends Member implements Serializable {
     private float x;
     private float y;
-    public float g_scores;
-    public float h_scores;
-    public float f_scores = 0;
+    public double g_scores;
+    public double h_scores;
+    public double f_scores = 0;
     private ArrayList<Edge> adjacencies;
     public Node parent;
     public boolean explored = false;
